Split the Application finishing logic from Support\Facades\Session to Support\Facades\App
