+++ system/Helpers/Url.php
@@ -12,7 +12,7 @@ use Helpers\Session;
 use Helpers\Inflector;
 
 use Support\Facades\Redirect;
-use Support\Facades\Session as SessionStore;
+use Support\Facades\App;
 
 
 /**
@@ -36,7 +36,7 @@ class Url
         $response = Redirect::to($url, $code);
 
         // Finish the Session (and send the Response).
-        SessionStore::finish($response);
+        App::finish($response);
 
         // Quit the Nova Framework's execution.
         exit();
